Refactoring:  refactoring of the middleware.test.js into refactorMiddlewareTest folder
