+++ test/middleware.test.js
@@ -62,755 +62,6 @@ describe('middleware', () => {
   }
 
   describe('basic', () => {
-    describe('should work', () => {
-      let compiler;
-      let codeLength;
-
-      const outputPath = path.resolve(__dirname, './outputs/basic');
-
-      beforeAll((done) => {
-        compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: outputPath,
-          },
-        });
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(() => {
-          compiler.hooks.afterCompile.tap('wdm-test', (params) => {
-            codeLength = params.assets['bundle.js'].source().length;
-            done();
-          });
-        });
-
-        instance.context.outputFileSystem.mkdirSync(outputPath, {
-          recursive: true,
-        });
-        instance.context.outputFileSystem.writeFileSync(
-          path.resolve(outputPath, 'image.svg'),
-          'svg image'
-        );
-        instance.context.outputFileSystem.mkdirSync(
-          path.resolve(outputPath, 'directory/nested-directory'),
-          { recursive: true }
-        );
-        instance.context.outputFileSystem.writeFileSync(
-          path.resolve(outputPath, 'directory/nested-directory/index.html'),
-          'My Index.'
-        );
-        instance.context.outputFileSystem.writeFileSync(
-          path.resolve(outputPath, 'throw-an-exception-on-readFileSync.txt'),
-          'exception'
-        );
-        instance.context.outputFileSystem.writeFileSync(
-          path.resolve(outputPath, 'unknown'),
-          'unknown'
-        );
-      });
-
-      afterAll(close);
-
-      it('should not find the bundle file on disk', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .expect('Content-Type', 'application/javascript; charset=utf-8')
-          .expect(200, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            expect(fs.existsSync(path.resolve(outputPath, 'bundle.js'))).toBe(
-              false
-            );
-
-            return done();
-          });
-      });
-
-      it('should return the "200" code for the "GET" request to the bundle file', (done) => {
-        const fileData = instance.context.outputFileSystem.readFileSync(
-          path.resolve(outputPath, 'bundle.js')
-        );
-
-        request(app)
-          .get('/bundle.js')
-          .expect('Content-Length', fileData.byteLength.toString())
-          .expect('Content-Type', 'application/javascript; charset=utf-8')
-          .expect(200, fileData.toString(), done);
-      });
-
-      it('should return the "200" code for the "HEAD" request to the bundle file', (done) => {
-        request(app)
-          .head('/bundle.js')
-          .expect(
-            'Content-Length',
-            instance.context.outputFileSystem
-              .readFileSync(path.resolve(outputPath, 'bundle.js'))
-              .byteLength.toString()
-          )
-          .expect('Content-Type', 'application/javascript; charset=utf-8')
-          // eslint-disable-next-line no-undefined
-          .expect(200, undefined, done);
-      });
-
-      it('should return the "404" code for the "POST" request to the bundle file', (done) => {
-        request(app).post('/bundle.js').expect(404, done);
-      });
-
-      it('should return the "200" code for the "GET" request to the "image.svg" file', (done) => {
-        const fileData = instance.context.outputFileSystem.readFileSync(
-          path.resolve(outputPath, 'image.svg')
-        );
-
-        request(app)
-          .get('/image.svg')
-          .expect('Content-Length', fileData.byteLength.toString())
-          .expect('Content-Type', 'image/svg+xml')
-          .expect(200, fileData, done);
-      });
-
-      it('should return the "200" code for the "GET" request to the directory', (done) => {
-        const fileData = fs.readFileSync(
-          path.resolve(__dirname, './fixtures/index.html')
-        );
-
-        request(app)
-          .get('/')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect('Content-Length', fileData.byteLength.toString())
-          .expect(200, fileData.toString(), done);
-      });
-
-      it('should return the "200" code for the "GET" request to the subdirectory with "index.html"', (done) => {
-        const fileData = instance.context.outputFileSystem.readFileSync(
-          path.resolve(outputPath, 'directory/nested-directory/index.html')
-        );
-
-        request(app)
-          .get('/directory/nested-directory/')
-          .expect('Content-Length', fileData.byteLength.toString())
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(200, fileData.toString(), done);
-      });
-
-      it('should return the "200" code for the "GET" request to the subdirectory with "index.html" without trailing slash', (done) => {
-        const fileData = instance.context.outputFileSystem.readFileSync(
-          path.resolve(outputPath, 'directory/nested-directory/index.html')
-        );
-
-        request(app)
-          .get('/directory/nested-directory')
-          .expect('Content-Length', fileData.byteLength.toString())
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(200, fileData.toString(), done);
-      });
-
-      it('should return the "200" code for the "GET" request to the subdirectory with "index.html"', (done) => {
-        const fileData = instance.context.outputFileSystem.readFileSync(
-          path.resolve(outputPath, 'directory/nested-directory/index.html')
-        );
-
-        request(app)
-          .get('/directory/nested-directory/index.html')
-          .expect('Content-Length', fileData.byteLength.toString())
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(200, fileData.toString(), done);
-      });
-
-      it('should return the "416" code for the "GET" request with the invalid range header', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .set('Range', 'bytes=9999999-')
-          .expect(416, done);
-      });
-
-      it('should return the "206" code for the "GET" request with the valid range header', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .set('Range', 'bytes=3000-3500')
-          .expect('Content-Length', '501')
-          .expect('Content-Range', `bytes 3000-3500/${codeLength}`)
-          .expect(206, done);
-      });
-
-      it('should return the "200" code for the "GET" request with malformed range header which is ignored', (done) => {
-        request(app).get('/bundle.js').set('Range', 'abc').expect(200, done);
-      });
-
-      it('should return the "200" code for the "GET" request with multiple range header which is ignored', (done) => {
-        request(app)
-          .get('/bundle.js')
-          .set('Range', 'bytes=3000-3100,3200-3300')
-          .expect(200, done);
-      });
-
-      it('should return the "404" code for the "GET" request with to the non-public path', (done) => {
-        request(app)
-          .get('/nonpublic/')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(404, done);
-      });
-
-      it('should return the "404" code for the "GET" request to the deleted file', (done) => {
-        const spy = jest
-          .spyOn(instance.context.outputFileSystem, 'readFileSync')
-          .mockImplementation(() => {
-            throw new Error('error');
-          });
-
-        request(app)
-          .get('/public/throw-an-exception-on-readFileSync.txt')
-          .expect(404, (error) => {
-            if (error) {
-              return done(error);
-            }
-
-            spy.mockRestore();
-
-            return done();
-          });
-      });
-
-      it('should return "200" code code for the "GET" request to the file without extension', (done) => {
-        const fileData = instance.context.outputFileSystem.readFileSync(
-          path.resolve(outputPath, 'unknown')
-        );
-
-        request(app)
-          .get('/unknown')
-          .expect('Content-Length', fileData.byteLength.toString())
-          .expect('Content-Type', 'application/octet-stream')
-          .expect(200, done);
-      });
-    });
-
-    describe('should not work with the broken "publicPath" option', () => {
-      let compiler;
-
-      const outputPath = path.resolve(__dirname, './outputs/basic');
-
-      beforeAll((done) => {
-        compiler = getCompiler({
-          ...webpackConfig,
-          output: {
-            filename: 'bundle.js',
-            path: outputPath,
-            publicPath: 'https://test:malfor%5Med@test.example.com',
-          },
-        });
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return the "400" code for the "GET" request to the bundle file', (done) => {
-        request(app).get('/bundle.js').expect(404, done);
-      });
-    });
-
-    describe('should work in multi-compiler mode', () => {
-      beforeAll((done) => {
-        const compiler = getCompiler(webpackMultiConfig);
-
-        instance = middleware(compiler);
-
-        app = express();
-        app.use(instance);
-
-        listen = listenShorthand(done);
-      });
-
-      afterAll(close);
-
-      it('should return "200" code for GET request to the bundle file for the first compiler', (done) => {
-        request(app).get('/static-one/bundle.js').expect(200, done);
-      });
-
-      it('should return "404" code for GET request to a non existing file for the first compiler', (done) => {
-        request(app).get('/static-one/invalid.js').expect(404, done);
-      });
-
-      it('should return "200" code for GET request to the "public" path for the first compiler', (done) => {
-        request(app)
-          .get('/static-one/')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(200, done);
-      });
-
-      it('should return "200" code for GET request to the "index" option for the first compiler', (done) => {
-        request(app)
-          .get('/static-one/index.html')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(200, done);
-      });
-
-      it('should return "200" code for GET request for the bundle file for the second compiler', (done) => {
-        request(app).get('/static-two/bundle.js').expect(200, done);
-      });
-
-      it('should return "404" code for GET request to a non existing file for the second compiler', (done) => {
-        request(app).get('/static-two/invalid.js').expect(404, done);
-      });
-
-      it('should return "404" code for GET request to the "public" path for the second compiler', (done) => {
-        request(app).get('/static-two/').expect(404, done);
-      });
-
-      it('should return "404" code for GET request to the "index" option for the second compiler', (done) => {
-        request(app).get('/static-two/index.html').expect(404, done);
-      });
-
-      it('should return "404" code for GET request to the non-public path', (done) => {
-        request(app)
-          .get('/static-three/')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(404, done);
-      });
-
-      it('should return "404" code for GET request to the non-public path', (done) => {
-        request(app)
-          .get('/static-three/invalid.js')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(404, done);
-      });
-
-      it('should return "404" code for GET request to the non-public path', (done) => {
-        request(app)
-          .get('/')
-          .expect('Content-Type', 'text/html; charset=utf-8')
-          .expect(404, done);
-      });
-    });
-
-    describe('should work with difference requests', () => {
-      const basicOutputPath = path.resolve(__dirname, './outputs/basic');
-      const fixtures = [
-        {
-          urls: [
-            {
-              value: 'bundle.js',
-              contentType: 'application/javascript; charset=utf-8',
-              code: 200,
-            },
-            {
-              value: '',
-              contentType: 'text/html; charset=utf-8',
-              code: 200,
-            },
-            {
-              value: 'index.html',
-              contentType: 'text/html; charset=utf-8',
-              code: 200,
-            },
-            {
-              value: 'invalid.js',
-              contentType: 'text/html; charset=utf-8',
-              code: 404,
-            },
-            {
-              value: 'complex',
-              contentType: 'text/html; charset=utf-8',
-              code: 404,
-            },
-            {
-              value: 'complex/invalid.js',
-              contentType: 'text/html; charset=utf-8',
-              code: 404,
-            },
-            {
-              value: 'complex/complex',
-              contentType: 'text/html; charset=utf-8',
-              code: 404,
-            },
-            {
-              value: 'complex/complex/invalid.js',
-              contentType: 'text/html; charset=utf-8',
-              code: 404,
-            },
-            {
-              value: '%',
-              contentType: 'text/html; charset=utf-8',
-              code: 404,
-            },
-          ],
-        },
-        {
-          file: 'config.json',
-          data: JSON.stringify({ foo: 'bar' }),
-          urls: [
-            {
-              value: 'config.json',
-              contentType: 'application/json; charset=utf-8',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: 'image.svg',
-          data: '<svg>SVG</svg>',
-          urls: [
-            {
-              value: 'image.svg',
-              contentType: 'image/svg+xml',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: 'foo.js',
-          data: 'console.log("foo");',
-          urls: [
-            {
-              value: 'foo.js',
-              contentType: 'application/javascript; charset=utf-8',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: '/complex/foo.js',
-          data: 'console.log("foo");',
-          urls: [
-            {
-              value: 'complex/foo.js',
-              contentType: 'application/javascript; charset=utf-8',
-              code: 200,
-            },
-            {
-              value: 'complex/./foo.js',
-              contentType: 'application/javascript; charset=utf-8',
-              code: 200,
-            },
-            {
-              value: 'complex/foo/../foo.js',
-              contentType: 'application/javascript; charset=utf-8',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: '/complex/complex/foo.js',
-          data: 'console.log("foo");',
-          urls: [
-            {
-              value: 'complex/complex/foo.js',
-              contentType: 'application/javascript; charset=utf-8',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: '/föö.js',
-          data: 'console.log("foo");',
-          urls: [
-            // Express encodes the URI component, so we do the same
-            {
-              value: 'f%C3%B6%C3%B6.js',
-              contentType: 'application/javascript; charset=utf-8',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: '/%foo%/%foo%.js',
-          data: 'console.log("foo");',
-          urls: [
-            // Filenames can contain characters not allowed in URIs
-            {
-              value: '%foo%/%foo%.js',
-              contentType: 'application/javascript; charset=utf-8',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: 'test.html',
-          data: '<div>test</div>',
-          urls: [
-            {
-              value: 'test.html?foo=bar',
-              contentType: 'text/html; charset=utf-8',
-              code: 200,
-            },
-            {
-              value: 'test.html?foo=bar#hash',
-              contentType: 'text/html; charset=utf-8',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: 'pathname with spaces.js',
-          data: 'console.log("foo");',
-          urls: [
-            {
-              value: 'pathname%20with%20spaces.js',
-              contentType: 'application/javascript; charset=utf-8',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: 'dirname with spaces/filename with spaces.js',
-          data: 'console.log("foo");',
-          urls: [
-            {
-              value: 'dirname%20with%20spaces/filename%20with%20spaces.js',
-              contentType: 'application/javascript; charset=utf-8',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: 'filename-name-with-dots/mono-v6.x.x',
-          data: 'content with .',
-          urls: [
-            {
-              value: 'filename-name-with-dots/mono-v6.x.x',
-              contentType: 'application/octet-stream',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: 'noextension',
-          data: 'noextension content',
-          urls: [
-            {
-              value: 'noextension',
-              contentType: 'application/octet-stream',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: '3dAr.usdz',
-          data: '3dAr.usdz content',
-          urls: [
-            {
-              value: '3dAr.usdz',
-              contentType: 'model/vnd.usdz+zip',
-              code: 200,
-            },
-          ],
-        },
-        {
-          file: 'hello.wasm',
-          data: 'hello.wasm content',
-          urls: [
-            {
-              value: 'hello.wasm',
-              contentType: 'application/wasm',
-              code: 200,
-            },
-          ],
-        },
-      ];
-
-      const configurations = [
-        {
-          output: { path: basicOutputPath, publicPath: '' },
-          publicPathForRequest: '/',
-        },
-        {
-          output: { path: path.join(basicOutputPath, 'dist'), publicPath: '' },
-          publicPathForRequest: '/',
-        },
-        {
-          output: { path: basicOutputPath, publicPath: '/' },
-          publicPathForRequest: '/',
-        },
-        {
-          output: { path: path.join(basicOutputPath, 'dist'), publicPath: '/' },
-          publicPathForRequest: '/',
-        },
-        {
-          output: { path: basicOutputPath, publicPath: '/static' },
-          publicPathForRequest: '/static/',
-        },
-        {
-          output: {
-            path: path.join(basicOutputPath, 'dist'),
-            publicPath: '/static',
-          },
-          publicPathForRequest: '/static/',
-        },
-        {
-          output: { path: basicOutputPath, publicPath: '/static/' },
-          publicPathForRequest: '/static/',
-        },
-        {
-          output: {
-            path: path.join(basicOutputPath, 'dist'),
-            publicPath: '/static/',
-          },
-          publicPathForRequest: '/static/',
-        },
-        {
-          output: {
-            path: path.join(basicOutputPath, 'dist/#leadinghash'),
-            publicPath: '/',
-          },
-          publicPathForRequest: '/',
-        },
-        {
-          output: {
-            path: basicOutputPath,
-            publicPath: 'http://127.0.0.1/',
-          },
-          publicPathForRequest: '/',
-        },
-        {
-          output: {
-            path: basicOutputPath,
-            publicPath: 'http://127.0.0.1:3000/',
-          },
-          publicPathForRequest: '/',
-        },
-        {
-          output: {
-            path: basicOutputPath,
-            publicPath: '//test.domain/',
-          },
-          publicPathForRequest: '/',
-        },
-      ];
-
-      const isWindows = process.platform === 'win32';
-
-      if (isWindows) {
-        fixtures.push(
-          {
-            file: 'windows.txt',
-            data: 'windows.txt content',
-            urls: [
-              {
-                value: 'windows.txt',
-                contentType: 'text/plain; charset=utf-8',
-                code: 200,
-              },
-            ],
-          },
-          {
-            file: 'windows 2.txt',
-            data: 'windows 2.txt content',
-            urls: [
-              {
-                value: 'windows%202.txt',
-                contentType: 'text/plain; charset=utf-8',
-                code: 200,
-              },
-            ],
-          },
-          {
-            file: 'test & test & %20.txt',
-            data: 'test & test & %20.txt content',
-            urls: [
-              {
-                value: 'test%20%26%20test%20%26%20%2520.txt',
-                contentType: 'text/plain; charset=utf-8',
-                code: 200,
-              },
-            ],
-          }
-        );
-
-        configurations.push(
-          {
-            output: {
-              path: path.join(basicOutputPath, 'my static'),
-              publicPath: '/static/',
-            },
-            publicPathForRequest: '/static/',
-          },
-          {
-            output: {
-              path: path.join(basicOutputPath, 'my%20static'),
-              publicPath: '/static/',
-            },
-            publicPathForRequest: '/static/',
-          },
-          {
-            output: {
-              path: path.join(basicOutputPath, 'my %20 static'),
-              publicPath: '/my%20static/',
-            },
-            publicPathForRequest: '/my%20static/',
-          }
-        );
-      }
-
-      for (const configuration of configurations) {
-        // eslint-disable-next-line no-loop-func
-        describe('should work handle requests', () => {
-          const { output, publicPathForRequest } = configuration;
-          const { path: outputPath, publicPath } = output;
-
-          let compiler;
-
-          beforeAll((done) => {
-            compiler = getCompiler({
-              ...webpackConfig,
-              output: {
-                filename: 'bundle.js',
-                path: outputPath,
-                publicPath,
-              },
-            });
-
-            instance = middleware(compiler);
-
-            app = express();
-            app.use(instance);
-
-            listen = listenShorthand(done);
-
-            const {
-              context: {
-                outputFileSystem: { mkdirSync, writeFileSync },
-              },
-            } = instance;
-
-            for (const { file, data } of fixtures) {
-              if (file) {
-                const fullPath = path.join(outputPath, file);
-
-                mkdirSync(path.dirname(fullPath), { recursive: true });
-                writeFileSync(fullPath, data);
-              }
-            }
-          });
-
-          afterAll(close);
-
-          for (const { data, urls } of fixtures) {
-            for (const { value, contentType, code } of urls) {
-              // eslint-disable-next-line no-loop-func
-              it(`should return the "${code}" code for the "GET" request for the "${value}" url`, (done) => {
-                request(app)
-                  .get(`${publicPathForRequest}${value}`)
-                  .expect('Content-Type', contentType)
-                  .expect('Content-Length', data ? String(data.length) : /\d+/)
-                  .expect(code, done);
-              });
-            }
-          }
-        });
-      }
-    });
-
     describe('should respect the value of the "Content-Type" header from other middleware', () => {
       beforeAll((done) => {
         const compiler = getCompiler(webpackConfig);
